# 8196106: Support nested infinite or recursive flat mapped streams