--- a/src/java.base/share/classes/java/util/stream/IntPipeline.java
+++ b/src/java.base/share/classes/java/util/stream/IntPipeline.java
@@ -24,8 +24,6 @@
  */
 package java.util.stream;
 
-import jdk.internal.vm.annotation.Stable;
-
 import java.util.IntSummaryStatistics;
 import java.util.Objects;
 import java.util.OptionalDouble;
@@ -43,7 +41,6 @@ import java.util.function.IntToDoubleFunction;
 import java.util.function.IntToLongFunction;
 import java.util.function.IntUnaryOperator;
 import java.util.function.ObjIntConsumer;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 /**
@@ -300,32 +297,43 @@ abstract class IntPipeline<E_IN>
                 StreamOpFlag.NOT_SORTED | StreamOpFlag.NOT_DISTINCT | StreamOpFlag.NOT_SIZED) {
             @Override
             Sink<Integer> opWrapSink(int flags, Sink<Integer> sink) {
-                class FlatMap extends Sink.ChainedInt<Integer> implements IntPredicate {
-                    @Stable boolean cancel;
+                final IntConsumer fastPath =
+                        isShortCircuitingPipeline()
+                                ? null
+                                : (sink instanceof IntConsumer ic)
+                                ? ic
+                                : sink::accept;
+                final class FlatMap implements Sink.OfInt, IntPredicate {
+                    boolean cancel;
 
-                    FlatMap() { super(sink); }
-
-                    @Override public void begin(long size) { downstream.begin(-1); }
+                    @Override public void begin(long size) { sink.begin(-1); }
+                    @Override public void end() { sink.end(); }
 
                     @Override
-                    public void accept(int input) {
-                        try (IntStream result = mapper.apply(input)) {
-                            if (result != null)
-                                result.sequential().allMatch(this);
+                    public void accept(int e) {
+                        try (final var result = mapper.apply(e)) {
+                            if (result != null) {
+                                if (fastPath == null)
+                                    result.sequential().allMatch(this);
+                                else
+                                    result.sequential().forEach(fastPath);
+                            }
                         }
                     }
 
                     @Override
                     public boolean cancellationRequested() {
-                        return cancel || (cancel |= downstream.cancellationRequested());
+                        return cancel || (cancel |= sink.cancellationRequested());
                     }
 
                     @Override
                     public boolean test(int output) {
-                        if (cancel)
+                        if (!cancel) {
+                            sink.accept(output);
+                            return !(cancel |= sink.cancellationRequested());
+                        } else {
                             return false;
-                        downstream.accept(output);
-                        return !cancellationRequested();
+                        }
                     }
                 }
                 return new FlatMap();