… src/java.base/share/classes/java/util/stream/ReferencePipeline.java
… test/micro/org/openjdk/bench/java/util/stream/ops/DoubleAccumulator.java
… test/micro/org/openjdk/bench/java/util/stream/ops/ref/FlatMap.java