This PR implements Gatherer-inspired encoding of `flatMap` that shows that it is both competitive performance-wise as well as improve correctness.

Below is the performance of `Stream::flatMap` (for reference types):

Before this PR (on my MacBook, aarch64):
```
Non-short-circuiting:

Benchmark                 (size)   Mode  Cnt        Score       Error  Units
FlatMap.par_array             10  thrpt   12   257582,480 ? 31360,242  ops/s
FlatMap.par_array            100  thrpt   12    55202,015 ? 14011,668  ops/s
FlatMap.par_array           1000  thrpt   12     6546,252 ?  3675,764  ops/s
FlatMap.par_doublestream      10  thrpt   12   267423,410 ? 37338,089  ops/s
FlatMap.par_doublestream     100  thrpt   12    27140,703 ?  4979,878  ops/s
FlatMap.par_doublestream    1000  thrpt   12     2978,178 ?  1890,250  ops/s
FlatMap.par_intstream         10  thrpt   12   268194,530 ? 37295,092  ops/s
FlatMap.par_intstream        100  thrpt   12    30897,034 ?  5388,245  ops/s
FlatMap.par_intstream       1000  thrpt   12     3969,043 ?  2494,641  ops/s
FlatMap.par_longstream        10  thrpt   12   279756,861 ? 19519,497  ops/s
FlatMap.par_longstream       100  thrpt   12    45733,955 ? 18385,144  ops/s
FlatMap.par_longstream      1000  thrpt   12     5115,615 ?  4147,237  ops/s
FlatMap.seq_array             10  thrpt   12  2937192,934 ? 58605,583  ops/s
FlatMap.seq_array            100  thrpt   12    41859,462 ?   139,021  ops/s
FlatMap.seq_array           1000  thrpt   12      442,677 ?     1,041  ops/s
FlatMap.seq_doublestream      10  thrpt   12  4972651,093 ? 35997,267  ops/s
FlatMap.seq_doublestream     100  thrpt   12    99265,005 ?   193,497  ops/s
FlatMap.seq_doublestream    1000  thrpt   12     1037,030 ?     3,254  ops/s
FlatMap.seq_intstream         10  thrpt   12  5133751,888 ? 23516,294  ops/s
FlatMap.seq_intstream        100  thrpt   12   145166,206 ?   399,263  ops/s
FlatMap.seq_intstream       1000  thrpt   12     1565,004 ?     3,207  ops/s
FlatMap.seq_longstream        10  thrpt   12  5047029,363 ? 24742,300  ops/s
FlatMap.seq_longstream       100  thrpt   12   233225,307 ?  7162,604  ops/s
FlatMap.seq_longstream      1000  thrpt   12     2999,926 ?     9,945  ops/s

// Short-circuiting:

Benchmark                   (size)   Mode  Cnt       Score       Error  Units
FlatMap.par_iterate_double      10  thrpt   12   46336,834 ?  6803,751  ops/s
FlatMap.par_iterate_double     100  thrpt   12   15365,884 ?  2750,656  ops/s
FlatMap.par_iterate_double    1000  thrpt   12    1677,618 ?   174,397  ops/s
FlatMap.par_iterate_int         10  thrpt   12   50275,901 ?  4321,430  ops/s
FlatMap.par_iterate_int        100  thrpt   12   16688,084 ?  2977,145  ops/s
FlatMap.par_iterate_int       1000  thrpt   12    1963,556 ?   384,075  ops/s
FlatMap.par_iterate_long        10  thrpt   12   50923,900 ?  3732,781  ops/s
FlatMap.par_iterate_long       100  thrpt   12   16660,676 ?  2086,627  ops/s
FlatMap.par_iterate_long      1000  thrpt   12    1949,490 ?   414,274  ops/s
FlatMap.par_iterate_ref         10  thrpt   12   26493,083 ?  5345,774  ops/s
FlatMap.par_iterate_ref        100  thrpt   12   10526,834 ?   165,435  ops/s
FlatMap.par_iterate_ref       1000  thrpt   12     923,592 ?     6,473  ops/s
FlatMap.seq_iterate_double      10  thrpt   12  457939,954 ? 21836,434  ops/s
FlatMap.seq_iterate_double     100  thrpt   12   10362,293 ?    79,699  ops/s
FlatMap.seq_iterate_double    1000  thrpt   12     101,664 ?     0,869  ops/s
FlatMap.seq_iterate_int         10  thrpt   12  438688,302 ?  8519,874  ops/s
FlatMap.seq_iterate_int        100  thrpt   12   10280,163 ?    94,136  ops/s
FlatMap.seq_iterate_int       1000  thrpt   12     100,687 ?     1,199  ops/s
FlatMap.seq_iterate_long        10  thrpt   12  437124,344 ? 39209,828  ops/s
FlatMap.seq_iterate_long       100  thrpt   12   10218,305 ?    80,571  ops/s
FlatMap.seq_iterate_long      1000  thrpt   12     100,560 ?     1,228  ops/s
FlatMap.seq_iterate_ref         10  thrpt   12  338190,816 ?  7213,549  ops/s
FlatMap.seq_iterate_ref        100  thrpt   12    8453,383 ?    44,707  ops/s
FlatMap.seq_iterate_ref       1000  thrpt   12      77,174 ?     0,701  ops/s
```

After this PR (on my MacBook, aarch64):

```
Non-short-circuiting:

Benchmark                 (size)   Mode  Cnt        Score       Error  Units
FlatMap.par_array             10  thrpt   12   309043,346 ? 22267,788  ops/s
FlatMap.par_array            100  thrpt   12    63773,183 ? 18772,809  ops/s
FlatMap.par_array           1000  thrpt   12     8454,577 ?   223,855  ops/s
FlatMap.par_doublestream      10  thrpt   12   287825,910 ? 40801,718  ops/s
FlatMap.par_doublestream     100  thrpt   12    29102,473 ?  4706,576  ops/s
FlatMap.par_doublestream    1000  thrpt   12     3966,528 ?  1915,358  ops/s
FlatMap.par_intstream         10  thrpt   12   291119,116 ? 32542,028  ops/s
FlatMap.par_intstream        100  thrpt   12    37945,217 ?  6986,065  ops/s
FlatMap.par_intstream       1000  thrpt   12     5004,935 ?  2325,120  ops/s
FlatMap.par_longstream        10  thrpt   12   306691,186 ? 30836,677  ops/s
FlatMap.par_longstream       100  thrpt   12    39204,098 ?  6159,602  ops/s
FlatMap.par_longstream      1000  thrpt   12     5197,154 ?  4293,582  ops/s
FlatMap.seq_array             10  thrpt   12  2852595,288 ? 19117,935  ops/s
FlatMap.seq_array            100  thrpt   12    41792,834 ?   125,631  ops/s
FlatMap.seq_array           1000  thrpt   12     1085,509 ?  1226,692  ops/s
FlatMap.seq_doublestream      10  thrpt   12  5405153,919 ? 25352,824  ops/s
FlatMap.seq_doublestream     100  thrpt   12    99297,136 ?   170,830  ops/s
FlatMap.seq_doublestream    1000  thrpt   12     1038,508 ?     1,066  ops/s
FlatMap.seq_intstream         10  thrpt   12  5558941,679 ? 25526,270  ops/s
FlatMap.seq_intstream        100  thrpt   12   145225,641 ?   334,990  ops/s
FlatMap.seq_intstream       1000  thrpt   12     1567,861 ?     2,790  ops/s
FlatMap.seq_longstream        10  thrpt   12  5592000,353 ? 20322,575  ops/s
FlatMap.seq_longstream       100  thrpt   12   241101,180 ?  9170,915  ops/s
FlatMap.seq_longstream      1000  thrpt   12     2132,999 ?  1589,173  ops/s

// Short-circuiting:

Benchmark                   (size)   Mode  Cnt        Score        Error  Units
FlatMap.par_iterate_double      10  thrpt   12    46354,848 ?   6372,823  ops/s
FlatMap.par_iterate_double     100  thrpt   12    15851,059 ?   2942,830  ops/s
FlatMap.par_iterate_double    1000  thrpt   12     1778,464 ?     27,042  ops/s
FlatMap.par_iterate_int         10  thrpt   12    51879,241 ?   6759,903  ops/s
FlatMap.par_iterate_int        100  thrpt   12    16458,214 ?   3206,146  ops/s
FlatMap.par_iterate_int       1000  thrpt   12     2421,501 ?      5,582  ops/s
FlatMap.par_iterate_long        10  thrpt   12    47412,155 ?   9957,661  ops/s
FlatMap.par_iterate_long       100  thrpt   12    17745,066 ?   3519,614  ops/s
FlatMap.par_iterate_long      1000  thrpt   12     2413,274 ?     27,152  ops/s
FlatMap.par_iterate_ref         10  thrpt   12    24754,662 ?   5144,389  ops/s
FlatMap.par_iterate_ref        100  thrpt   12    10535,374 ?    273,802  ops/s
FlatMap.par_iterate_ref       1000  thrpt   12      990,926 ?    153,386  ops/s
FlatMap.seq_iterate_double      10  thrpt   12   941441,571 ?  72849,560  ops/s
FlatMap.seq_iterate_double     100  thrpt   12    26116,346 ?    225,451  ops/s
FlatMap.seq_iterate_double    1000  thrpt   12      285,873 ?      1,924  ops/s
FlatMap.seq_iterate_int         10  thrpt   12  1097970,188 ? 365428,186  ops/s
FlatMap.seq_iterate_int        100  thrpt   12    28136,050 ?    136,249  ops/s
FlatMap.seq_iterate_int       1000  thrpt   12      316,555 ?      0,687  ops/s
FlatMap.seq_iterate_long        10  thrpt   12  1021716,173 ?   8729,180  ops/s
FlatMap.seq_iterate_long       100  thrpt   12    28290,057 ?    264,404  ops/s
FlatMap.seq_iterate_long      1000  thrpt   12      316,620 ?      1,906  ops/s
FlatMap.seq_iterate_ref         10  thrpt   12   820498,644 ? 167233,794  ops/s
FlatMap.seq_iterate_ref        100  thrpt   12    13583,900 ?    116,079  ops/s
FlatMap.seq_iterate_ref       1000  thrpt   12      158,580 ?      1,977  ops/s
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8196106](https://bugs.openjdk.org/browse/JDK-8196106): Support nested infinite or recursive flat mapped streams (**Bug** - P4)


### Reviewers
 * [Paul Sandoz](https://openjdk.org/census#psandoz) (@PaulSandoz - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/18625/head:pull/18625` \
`$ git checkout pull/18625`

Update a local copy of the PR: \
`$ git checkout pull/18625` \
`$ git pull https://git.openjdk.org/jdk.git pull/18625/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 18625`

View PR using the GUI difftool: \
`$ git pr show -t 18625`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/18625.diff">https://git.openjdk.org/jdk/pull/18625.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/18625#issuecomment-2044620346)