@viktorklang-ora 2024-04-12
@PaulSandoz @AlanBateman This should be ready for review now, if you find some time :)