@forax 2024-04-13
Hello, i will not pretend I fully understand the code, but i think it's important to keep a constistent code style.
(and no I do not want to start an holy war about the code style, I just think we should keep it consistent)